enhance: move the implementation of S3Client out of cpp
This commit moves the implementation of S3Client out of
`multi_part_upload_s3_fs` (because it has become overly
bloated). After the storage team decided not to remove
the `multi_part_upload_s3_fs` and to maintain a private
S3 filesystem(port from arrow), then we should reoriginal
arrow implementation should be split out into different
files.

However, this does not mean that internal APIs can directly
call S3Client. Except for special cases, we should still
uniformly use the filesystem's API.
