+++ cpp/include/milvus-storage/filesystem/s3/multi_part_upload_s3_fs.h
@@ -28,53 +28,12 @@
 #include "arrow/io/interfaces.h"
 #include "milvus-storage/common/constants.h"
 #include "milvus-storage/filesystem/s3/s3_options.h"
+#include "milvus-storage/filesystem/s3/s3_client.h"
 
 using ::arrow::fs::FileInfo;
 using ::arrow::fs::FileInfoGenerator;
 
 namespace milvus_storage {
-
-class S3ClientMetrics {
-  public:
-  S3ClientMetrics() = default;
-  ~S3ClientMetrics() = default;
-
-  void IncrementMultiPartUploadCreated() { multi_part_upload_created.fetch_add(1, std::memory_order_relaxed); }
-  void IncrementMultiPartUploadFinished() { multi_part_upload_finished.fetch_add(1, std::memory_order_relaxed); }
-  void IncrementUploadBytes(int64_t bytes) { upload_bytes.fetch_add(bytes, std::memory_order_relaxed); }
-  void IncrementDownloadBytes(int64_t bytes) { download_bytes.fetch_add(bytes, std::memory_order_relaxed); }
-  void IncrementUploadCount() { upload_count.fetch_add(1, std::memory_order_relaxed); }
-  void IncrementDownloadCount() { download_count.fetch_add(1, std::memory_order_relaxed); }
-  void IncrementFailedCount() { failed_count.fetch_add(1, std::memory_order_relaxed); }
-
-  int64_t GetMultiPartUploadCreated() const { return multi_part_upload_created.load(std::memory_order_relaxed); }
-  int64_t GetMultiPartUploadFinished() const { return multi_part_upload_finished.load(std::memory_order_relaxed); }
-  int64_t GetUploadCount() const { return upload_count.load(std::memory_order_relaxed); }
-  int64_t GetDownloadCount() const { return download_count.load(std::memory_order_relaxed); }
-  int64_t GetUploadBytes() const { return upload_bytes.load(std::memory_order_relaxed); }
-  int64_t GetDownloadBytes() const { return download_bytes.load(std::memory_order_relaxed); }
-  int64_t GetFailedCount() const { return failed_count.load(std::memory_order_relaxed); }
-
-  void Reset() {
-    multi_part_upload_created.store(0, std::memory_order_relaxed);
-    multi_part_upload_finished.store(0, std::memory_order_relaxed);
-    upload_count.store(0, std::memory_order_relaxed);
-    download_count.store(0, std::memory_order_relaxed);
-    upload_bytes.store(0, std::memory_order_relaxed);
-    download_bytes.store(0, std::memory_order_relaxed);
-    failed_count.store(0, std::memory_order_relaxed);
-  }
-
-  private:
-  std::atomic<int64_t> multi_part_upload_created{0};
-  std::atomic<int64_t> multi_part_upload_finished{0};
-  std::atomic<int64_t> upload_count{0};
-  std::atomic<int64_t> download_count{0};
-  std::atomic<int64_t> failed_count{0};
-
-  std::atomic<int64_t> upload_bytes{0};
-  std::atomic<int64_t> download_bytes{0};
-};
 
 class MultiPartUploadS3FS : public arrow::fs::FileSystem {
   public:
