+++ cpp/include/milvus-storage/filesystem/s3/provider/TencentCloudSTSClient.h
