+++ cpp/src/filesystem/s3/multi_part_upload_s3_fs.cpp
@@ -18,7 +18,7 @@
 
 #include "milvus-storage/common/path_util.h"
 #include "milvus-storage/filesystem/io/io_util.h"
-
+#include "milvus-storage/filesystem/s3/s3_client.h"
 #include "arrow/util/async_generator.h"
 #include "arrow/util/logging.h"
 #include "arrow/buffer.h"
@@ -107,7 +107,7 @@ using ::milvus_storage::fs::internal::S3Backend;
 namespace S3Model = Aws::S3::Model;
 
 namespace milvus_storage {
-
+using namespace fs::internal;
 // -----------------------------------------------------------------------
 // MultiPartUploadS3FS implementation
 
@@ -134,14 +134,6 @@ bool IsDirectory(std::string_view key, const S3Model::HeadObjectResult& result) 
   // Otherwise, it's a regular file.
   return false;
 }
-
-inline Aws::String ToAwsString(const std::string& s) {
-  // Direct construction of Aws::String from std::string doesn't work because
-  // it uses a specific Allocator class.
-  return Aws::String(s.begin(), s.end());
-}
-
-inline std::string_view FromAwsString(const Aws::String& s) { return {s.data(), s.length()}; }
 
 template <typename ObjectRequest>
 struct ObjectMetadataSetter {
@@ -276,12 +268,10 @@ arrow::Status ValidateFilePath(const S3Path& path) {
   return arrow::Status::OK();
 };
 
-arrow::Status ErrorS3Finalized() { return arrow::Status::Invalid("S3 subsystem is finalized"); }
-
 arrow::Status CheckS3Initialized() {
   if (!IsS3Initialized()) {
     if (IsS3Finalized()) {
-      return ErrorS3Finalized();
+      return arrow::Status::Invalid("S3 subsystem is finalized");
     }
     return arrow::Status::Invalid(
         "S3 subsystem is not initialized; please call InitializeS3() "
@@ -289,382 +279,6 @@ arrow::Status CheckS3Initialized() {
   }
   return arrow::Status::OK();
 };
-
-class WrappedRetryStrategy : public Aws::Client::RetryStrategy {
-  public:
-  explicit WrappedRetryStrategy(const std::shared_ptr<arrow::fs::S3RetryStrategy>& s3_retry_strategy)
-      : s3_retry_strategy_(s3_retry_strategy) {}
-
-  bool ShouldRetry(const Aws::Client::AWSError<Aws::Client::CoreErrors>& error,
-                   long attempted_retries) const override {  // NOLINT runtime/int
-    arrow::fs::S3RetryStrategy::AWSErrorDetail detail = ErrorToDetail(error);
-    return s3_retry_strategy_->ShouldRetry(detail, static_cast<int64_t>(attempted_retries));
-  }
-
-  long CalculateDelayBeforeNextRetry(  // NOLINT runtime/int
-      const Aws::Client::AWSError<Aws::Client::CoreErrors>& error,
-      long attempted_retries) const override {  // NOLINT runtime/int
-    arrow::fs::S3RetryStrategy::AWSErrorDetail detail = ErrorToDetail(error);
-    return static_cast<long>(  // NOLINT runtime/int
-        s3_retry_strategy_->CalculateDelayBeforeNextRetry(detail, static_cast<int64_t>(attempted_retries)));
-  }
-
-  private:
-  template <typename ErrorType>
-  static arrow::fs::S3RetryStrategy::AWSErrorDetail ErrorToDetail(const Aws::Client::AWSError<ErrorType>& error) {
-    arrow::fs::S3RetryStrategy::AWSErrorDetail detail;
-    detail.error_type = static_cast<int>(error.GetErrorType());
-    detail.message = std::string(FromAwsString(error.GetMessage()));
-    detail.exception_name = std::string(FromAwsString(error.GetExceptionName()));
-    detail.should_retry = error.ShouldRetry();
-    return detail;
-  }
-
-  std::shared_ptr<arrow::fs::S3RetryStrategy> s3_retry_strategy_;
-};
-
-class S3Client : public Aws::S3::S3Client {
-  public:
-  using Aws::S3::S3Client::S3Client;
-
-  static inline constexpr auto kBucketRegionHeaderName = "x-amz-bucket-region";
-
-  std::string GetBucketRegionFromHeaders(const Aws::Http::HeaderValueCollection& headers) {
-    const auto it = headers.find(ToAwsString(kBucketRegionHeaderName));
-    if (it != headers.end()) {
-      return std::string(FromAwsString(it->second));
-    }
-    return std::string();
-  }
-
-  template <typename ErrorType>
-  arrow::Result<std::string> GetBucketRegionFromError(const std::string& bucket,
-                                                      const Aws::Client::AWSError<ErrorType>& error) {
-    std::string region = GetBucketRegionFromHeaders(error.GetResponseHeaders());
-    if (!region.empty()) {
-      return region;
-    } else if (error.GetResponseCode() == Aws::Http::HttpResponseCode::NOT_FOUND) {
-      return arrow::Status::IOError("Bucket '", bucket, "' not found");
-    } else {
-      return arrow::Status::IOError("When resolving region for bucket: ", bucket);
-    }
-  }
-
-  arrow::Result<std::string> GetBucketRegion(const std::string& bucket, const S3Model::HeadBucketRequest& request) {
-    auto uri = GeneratePresignedUrl(request.GetBucket(),
-                                    /*key=*/"", Aws::Http::HttpMethod::HTTP_HEAD);
-    // NOTE: The signer region argument isn't passed here, as there's no easy
-    // way of computing it (the relevant method is private).
-    auto outcome = MakeRequest(uri, request, Aws::Http::HttpMethod::HTTP_HEAD, Aws::Auth::SIGV4_SIGNER);
-    if (!outcome.IsSuccess()) {
-      return GetBucketRegionFromError(bucket, outcome.GetError());
-    }
-    std::string region = GetBucketRegionFromHeaders(outcome.GetResult().GetHeaderValueCollection());
-    if (!region.empty()) {
-      return region;
-    } else if (outcome.GetResult().GetResponseCode() == Aws::Http::HttpResponseCode::NOT_FOUND) {
-      return arrow::Status::IOError("Bucket '", request.GetBucket(), "' not found");
-    } else {
-      return arrow::Status::IOError("When resolving region for bucket '", request.GetBucket(),
-                                    "': missing 'x-amz-bucket-region' header in response");
-    }
-  }
-
-  arrow::Result<std::string> GetBucketRegion(const std::string& bucket) {
-    S3Model::HeadBucketRequest req;
-    req.SetBucket(ToAwsString(bucket));
-    return GetBucketRegion(bucket, req);
-  }
-
-  S3Model::CompleteMultipartUploadOutcome CompleteMultipartUploadWithErrorFixup(
-      S3Model::CompleteMultipartUploadRequest&& request) const {
-    // CompletedMultipartUpload can return a 200 OK response with an error
-    // encoded in the response body, in which case we should either retry
-    // or propagate the error to the user (see
-    // https://docs.aws.amazon.com/AmazonS3/latest/API/API_CompleteMultipartUpload.html).
-    //
-    // Unfortunately the AWS SDK doesn't detect such situations but lets them
-    // return successfully (see https://github.com/aws/aws-sdk-cpp/issues/658).
-    //
-    // We work around the issue by registering a DataReceivedEventHandler
-    // which parses the XML response for embedded errors.
-
-    std::optional<AWSError<Aws::Client::CoreErrors>> aws_error;
-    metrics_->IncrementMultiPartUploadFinished();
-
-    auto handler = [&](const Aws::Http::HttpRequest* http_req, Aws::Http::HttpResponse* http_resp,
-                       long long) {  // NOLINT runtime/int
-      auto& stream = http_resp->GetResponseBody();
-      const auto pos = stream.tellg();
-      const auto doc = Aws::Utils::Xml::XmlDocument::CreateFromXmlStream(stream);
-      // Rewind stream for later
-      stream.clear();
-      stream.seekg(pos);
-
-      if (doc.WasParseSuccessful()) {
-        auto root = doc.GetRootElement();
-        if (!root.IsNull()) {
-          // Detect something that looks like an abnormal CompletedMultipartUpload
-          // response.
-          if (root.GetName() != "CompleteMultipartUploadResult" || !root.FirstChild("Error").IsNull() ||
-              !root.FirstChild("Errors").IsNull()) {
-            // Make sure the error marshaller doesn't see a 200 OK
-            http_resp->SetResponseCode(Aws::Http::HttpResponseCode::INTERNAL_SERVER_ERROR);
-            aws_error = GetErrorMarshaller()->Marshall(*http_resp);
-            // Rewind stream for later
-            stream.clear();
-            stream.seekg(pos);
-          }
-        }
-      }
-    };
-
-    request.SetDataReceivedEventHandler(std::move(handler));
-
-    // We don't have access to the configured AWS retry strategy
-    // (m_retryStrategy is a private member of AwsClient), so don't use that.
-    std::unique_ptr<Aws::Client::RetryStrategy> retry_strategy;
-    if (s3_retry_strategy_) {
-      retry_strategy.reset(new WrappedRetryStrategy(s3_retry_strategy_));
-    } else {
-      // Note that DefaultRetryStrategy, unlike StandardRetryStrategy,
-      // has empty definitions for RequestBookkeeping() and GetSendToken(),
-      // which simplifies the code below.
-      retry_strategy.reset(new Aws::Client::DefaultRetryStrategy());
-    }
-
-    for (int32_t retries = 0;; retries++) {
-      aws_error.reset();
-      auto outcome = Aws::S3::S3Client::S3Client::CompleteMultipartUpload(request);
-      if (!outcome.IsSuccess()) {
-        // Error returned in HTTP headers (or client failure)
-        return outcome;
-      }
-      if (!aws_error.has_value()) {
-        // Genuinely successful outcome
-        return outcome;
-      }
-
-      const bool should_retry = retry_strategy->ShouldRetry(*aws_error, retries);
-
-      ARROW_LOG(WARNING) << "CompletedMultipartUpload got error embedded in a 200 OK response: "
-                         << aws_error->GetExceptionName() << " (\"" << aws_error->GetMessage()
-                         << "\"), retry = " << should_retry;
-
-      if (!should_retry) {
-        break;
-      }
-      const auto delay = std::chrono::milliseconds(retry_strategy->CalculateDelayBeforeNextRetry(*aws_error, retries));
-      std::this_thread::sleep_for(delay);
-    }
-
-    DCHECK(aws_error.has_value());
-    auto s3_error = AWSError<S3Errors>(std::move(aws_error).value());
-    return S3Model::CompleteMultipartUploadOutcome(std::move(s3_error));
-  }
-
-  std::shared_ptr<arrow::fs::S3RetryStrategy> s3_retry_strategy_;
-
-  // Metrics related functions
-  S3Model::CreateMultipartUploadOutcome CreateMultipartUpload(
-      const Aws::S3::Model::CreateMultipartUploadRequest& request) const override {
-    metrics_->IncrementMultiPartUploadCreated();
-    auto outcome = Aws::S3::S3Client::CreateMultipartUpload(request);
-    if (!outcome.IsSuccess()) {
-      metrics_->IncrementFailedCount();
-    }
-    return outcome;
-  }
-
-  S3Model::UploadPartOutcome UploadPart(const Aws::S3::Model::UploadPartRequest& request) const override {
-    metrics_->IncrementUploadCount();
-    auto outcome = Aws::S3::S3Client::UploadPart(request);
-    if (!outcome.IsSuccess()) {
-      metrics_->IncrementFailedCount();
-    } else {
-      metrics_->IncrementUploadBytes(request.GetContentLength());
-    }
-    return outcome;
-  }
-
-  S3Model::PutObjectOutcome PutObject(const Aws::S3::Model::PutObjectRequest& request) const override {
-    metrics_->IncrementUploadCount();
-    auto outcome = Aws::S3::S3Client::PutObject(request);
-    if (!outcome.IsSuccess()) {
-      metrics_->IncrementFailedCount();
-    } else {
-      metrics_->IncrementUploadBytes(request.GetContentLength());
-    }
-    return outcome;
-  }
-
-  S3Model::GetObjectOutcome GetObject(const Aws::S3::Model::GetObjectRequest& request) const override {
-    metrics_->IncrementDownloadCount();
-    auto outcome = Aws::S3::S3Client::GetObject(request);
-    if (!outcome.IsSuccess()) {
-      metrics_->IncrementFailedCount();
-    } else {
-      metrics_->IncrementDownloadBytes(outcome.GetResult().GetContentLength());
-    }
-    return outcome;
-  }
-
-  std::shared_ptr<S3ClientMetrics> GetMetrics() const { return metrics_; }
-
-  private:
-  std::shared_ptr<S3ClientMetrics> metrics_ = std::make_shared<S3ClientMetrics>();
-};
-
-class S3ClientFinalizer;
-
-class S3ClientLock {
-  public:
-  S3Client* get() { return client_.get(); }
-  S3Client* operator->() { return client_.get(); }
-
-  // Move this S3ClientLock into a temporary instance
-  //
-  // It is counter-intuitive, but lock ordering issues can happen even
-  // with a shared mutex locked in shared mode.
-  // The reason is that locking again in shared mode can block while
-  // there are threads waiting to take the lock in exclusive mode.
-  // Therefore, we should avoid obtaining the S3ClientLock when
-  // we already have it locked.
-  //
-  // This methods helps by moving the S3ClientLock into a temporary
-  // that is immediately destroyed so the lock will be released as
-  // soon as we are done making the call to the underlying client.
-  //
-  // (see GH-36523)
-  S3ClientLock Move() { return std::move(*this); }
-
-  protected:
-  friend class S3ClientHolder;
-
-  // Locks the finalizer until the S3ClientLock gets out of scope.
-  std::shared_lock<std::shared_mutex> lock_;
-  std::shared_ptr<S3Client> client_;
-};
-
-class S3ClientHolder {
-  public:
-  /// \brief Return a RAII guard guaranteeing a S3Client is safe for use
-  ///
-  /// S3 finalization will be deferred until the returned S3ClientLock
-  /// goes out of scope.
-  /// An error is returned if S3 is already finalized.
-  arrow::Result<S3ClientLock> Lock();
-
-  S3ClientHolder(std::weak_ptr<S3ClientFinalizer> finalizer, std::shared_ptr<S3Client> client)
-      : finalizer_(std::move(finalizer)), client_(std::move(client)) {}
-
-  void Finalize();
-
-  protected:
-  std::mutex mutex_;
-  std::weak_ptr<S3ClientFinalizer> finalizer_;
-  std::shared_ptr<S3Client> client_;
-};
-
-class S3ClientFinalizer : public std::enable_shared_from_this<S3ClientFinalizer> {
-  using ClientHolderList = std::vector<std::weak_ptr<S3ClientHolder>>;
-
-  public:
-  arrow::Result<std::shared_ptr<S3ClientHolder>> AddClient(std::shared_ptr<S3Client> client) {
-    std::unique_lock lock(mutex_);
-    if (finalized_) {
-      return ErrorS3Finalized();
-    }
-
-    auto holder = std::make_shared<S3ClientHolder>(shared_from_this(), std::move(client));
-
-    // Remove expired entries before adding new one
-    auto end = std::remove_if(holders_.begin(), holders_.end(),
-                              [](std::weak_ptr<S3ClientHolder> holder) { return holder.expired(); });
-    holders_.erase(end, holders_.end());
-    holders_.emplace_back(holder);
-    return holder;
-  }
-
-  void Finalize() {
-    std::unique_lock lock(mutex_);
-    finalized_ = true;
-
-    ClientHolderList finalizing = std::move(holders_);
-    lock.unlock();  // avoid lock ordering issue with S3ClientHolder::Finalize
-
-    // Finalize all client holders, such that no S3Client remains alive
-    // after this.
-    for (auto&& weak_holder : finalizing) {
-      auto holder = weak_holder.lock();
-      if (holder) {
-        holder->Finalize();
-      }
-    }
-  }
-
-  auto LockShared() { return std::shared_lock(mutex_); }
-
-  protected:
-  friend class S3ClientHolder;
-
-  std::shared_mutex mutex_;
-  ClientHolderList holders_;
-  bool finalized_ = false;
-};
-
-arrow::Result<S3ClientLock> S3ClientHolder::Lock() {
-  std::shared_ptr<S3ClientFinalizer> finalizer;
-  std::shared_ptr<S3Client> client;
-  {
-    std::unique_lock lock(mutex_);
-    finalizer = finalizer_.lock();
-    client = client_;
-  }
-  // Do not hold mutex while taking finalizer lock below.
-  //
-  // Acquiring a shared_mutex in shared mode may block even if not already
-  // acquired in exclusive mode, because of pending writers:
-  // https://github.com/google/sanitizers/issues/1668#issuecomment-1624985664
-  // """It is implementation-defined whether the calling thread acquires
-  // the lock when a writer does not hold the lock and there are writers
-  // blocked on the lock""".
-  //
-  // Therefore, we want to avoid potential lock ordering issues
-  // even when a shared lock is involved (GH-36523).
-  if (!finalizer) {
-    return ErrorS3Finalized();
-  }
-
-  S3ClientLock client_lock;
-  // Lock the finalizer before examining it
-  client_lock.lock_ = finalizer->LockShared();
-  if (finalizer->finalized_) {
-    return ErrorS3Finalized();
-  }
-  // (the client can be cleared only if finalizer->finalized_ is true)
-  DCHECK(client) << "inconsistent S3ClientHolder";
-  client_lock.client_ = std::move(client);
-  return client_lock;
-}
-
-void S3ClientHolder::Finalize() {
-  std::shared_ptr<S3Client> client;
-  {
-    std::unique_lock lock(mutex_);
-    client = std::move(client_);
-  }
-  // Do not hold mutex while ~S3Client potentially runs
-}
-
-std::shared_ptr<S3ClientFinalizer> GetClientFinalizer() {
-  static auto finalizer = std::make_shared<S3ClientFinalizer>();
-  return finalizer;
-}
-
-arrow::Result<std::shared_ptr<S3ClientHolder>> GetClientHolder(std::shared_ptr<S3Client> client) {
-  return GetClientFinalizer()->AddClient(std::move(client));
-}
 
 template <typename ObjectRequest>
 arrow::Status SetObjectMetadata(const std::shared_ptr<const arrow::KeyValueMetadata>& metadata, ObjectRequest* req) {
@@ -689,108 +303,6 @@ class StringViewStream : Aws::Utils::Stream::PreallocatedStreamBuf, public std::
       : Aws::Utils::Stream::PreallocatedStreamBuf(reinterpret_cast<unsigned char*>(const_cast<void*>(data)),
                                                   static_cast<size_t>(nbytes)),
         std::iostream(this) {}
-};
-
-class ClientBuilder {
-  public:
-  explicit ClientBuilder(S3Options options) : options_(std::move(options)) {}
-
-  const Aws::Client::ClientConfiguration& config() const { return client_config_; }
-
-  Aws::Client::ClientConfiguration* mutable_config() { return &client_config_; }
-
-  arrow::Result<std::shared_ptr<S3ClientHolder>> BuildClient(
-      std::optional<arrow::io::IOContext> io_context = std::nullopt) {
-    credentials_provider_ = options_.credentials_provider;
-    if (!options_.region.empty()) {
-      client_config_.region = ToAwsString(options_.region);
-    }
-    if (options_.request_timeout > 0) {
-      // Use ceil() to avoid setting it to 0 as that probably means no timeout.
-      client_config_.requestTimeoutMs = static_cast<long>(ceil(options_.request_timeout * 1000));  // NOLINT runtime/int
-    }
-    if (options_.connect_timeout > 0) {
-      client_config_.connectTimeoutMs = static_cast<long>(ceil(options_.connect_timeout * 1000));  // NOLINT runtime/int
-    }
-
-    client_config_.endpointOverride = ToAwsString(options_.endpoint_override);
-    if (options_.scheme == "http") {
-      client_config_.scheme = Aws::Http::Scheme::HTTP;
-      client_config_.verifySSL = false;
-    } else if (options_.scheme == "https") {
-      client_config_.scheme = Aws::Http::Scheme::HTTPS;
-      client_config_.verifySSL = true;
-    } else {
-      return arrow::Status::Invalid("Invalid S3 connection scheme '", options_.scheme, "'");
-    }
-    if (options_.retry_strategy) {
-      client_config_.retryStrategy = std::make_shared<WrappedRetryStrategy>(options_.retry_strategy);
-    } else {
-      client_config_.retryStrategy = std::make_shared<ConnectRetryStrategy>();
-    }
-    if (!arrow::fs::internal::global_options.tls_ca_file_path.empty()) {
-      client_config_.caFile = ToAwsString(arrow::fs::internal::global_options.tls_ca_file_path);
-      client_config_.verifySSL = false;
-    }
-    if (!arrow::fs::internal::global_options.tls_ca_dir_path.empty()) {
-      client_config_.caPath = ToAwsString(arrow::fs::internal::global_options.tls_ca_dir_path);
-    }
-
-    // Set proxy options if provided
-    if (!options_.proxy_options.scheme.empty()) {
-      if (options_.proxy_options.scheme == "http") {
-        client_config_.proxyScheme = Aws::Http::Scheme::HTTP;
-      } else if (options_.proxy_options.scheme == "https") {
-        client_config_.proxyScheme = Aws::Http::Scheme::HTTPS;
-      } else {
-        return arrow::Status::Invalid("Invalid proxy connection scheme '", options_.proxy_options.scheme, "'");
-      }
-    }
-    if (!options_.proxy_options.host.empty()) {
-      client_config_.proxyHost = ToAwsString(options_.proxy_options.host);
-    }
-    if (options_.proxy_options.port != -1) {
-      client_config_.proxyPort = options_.proxy_options.port;
-    }
-    if (!options_.proxy_options.username.empty()) {
-      client_config_.proxyUserName = ToAwsString(options_.proxy_options.username);
-    }
-    if (!options_.proxy_options.password.empty()) {
-      client_config_.proxyPassword = ToAwsString(options_.proxy_options.password);
-    }
-
-    if (io_context) {
-      // TODO: Once ARROW-15035 is done we can get rid of the "at least 25" fallback
-      client_config_.maxConnections = std::max(io_context->executor()->GetCapacity(), 25);
-    }
-
-    client_config_.maxConnections = std::max(client_config_.maxConnections, options_.max_connections);
-
-    const bool use_virtual_addressing = options_.endpoint_override.empty() || options_.force_virtual_addressing;
-
-#ifdef ARROW_S3_HAS_S3CLIENT_CONFIGURATION
-    client_config_.useVirtualAddressing = use_virtual_addressing;
-    auto endpoint_provider = EndpointProviderCache::Instance()->Lookup(client_config_);
-    auto client = std::make_shared<S3Client>(credentials_provider_, endpoint_provider, client_config_);
-#else
-    auto client =
-        std::make_shared<S3Client>(credentials_provider_, client_config_,
-                                   Aws::Client::AWSAuthV4Signer::PayloadSigningPolicy::Never, use_virtual_addressing);
-#endif
-    client->s3_retry_strategy_ = options_.retry_strategy;
-    return GetClientHolder(std::move(client));
-  }
-
-  const S3Options& options() const { return options_; }
-
-  protected:
-  S3Options options_;
-#ifdef ARROW_S3_HAS_S3CLIENT_CONFIGURATION
-  Aws::S3::S3ClientConfiguration client_config_;
-#else
-  Aws::Client::ClientConfiguration client_config_;
-#endif
-  std::shared_ptr<Aws::Auth::AWSCredentialsProvider> credentials_provider_;
 };
 
 std::string FormatRange(int64_t start, int64_t length) {
